login added in competitions

+++ controllers/parameters.ts
@@ -3,11 +3,18 @@ import queryDatabase from "../database/connection";
 
 export const getAllParameters = async (req: Request, res: Response) => {
   try {
+    const { competition_id } = req.headers;
+    if (!competition_id) {
+      return res
+        .status(500)
+        .json({ message: "Competition id not present", success: false });
+    }
     const query = `
       SELECT *
-      FROM parameters;
+      FROM parameters
+      WHERE competition_id = $1;
     `;
-    const parameters = await queryDatabase(query);
+    const parameters = await queryDatabase(query, [competition_id]);
     res.status(200).json({ parameters, success: true });
   } catch (error) {
     res.status(500).json({
@@ -20,14 +27,20 @@ export const getAllParameters = async (req: Request, res: Response) => {
 export const getSingleParameter = async (req: Request, res: Response) => {
   try {
     const parameterId = req.params.id;
+    const { competition_id } = req.headers;
+    if (!competition_id) {
+      return res
+        .status(500)
+        .json({ message: "Competition id not present", success: false });
+    }
 
     const query = `
       SELECT *
       FROM parameters
-      WHERE parameter_id = $1;
+      WHERE parameter_id = $1 AND competition_id = $2;
     `;
 
-    const parameter = await queryDatabase(query, [parameterId]);
+    const parameter = await queryDatabase(query, [parameterId, competition_id]);
 
     if (parameter.length === 0) {
       res.status(404).json({ message: "Parameter not found", success: false });
@@ -44,14 +57,20 @@ export const getSingleParameter = async (req: Request, res: Response) => {
 export const createParameter = async (req: Request, res: Response) => {
   try {
     const { parameter_name, parameter_description } = req.body;
+    const { competition_id } = req.headers;
+    if (!competition_id) {
+      return res
+        .status(500)
+        .json({ message: "Competition id not present", success: false });
+    }
 
     const query = `
-      INSERT INTO parameters ( parameter_name, parameter_description)
-      VALUES ($1, $2)
+      INSERT INTO parameters ( parameter_name, parameter_description, competition_id)
+      VALUES ($1, $2, $3)
       RETURNING *;
     `;
 
-    const values = [parameter_name, parameter_description];
+    const values = [parameter_name, parameter_description, competition_id];
 
     const newParameter = await queryDatabase(query, values);
 
@@ -68,16 +87,27 @@ export const updateParameter = async (req: Request, res: Response) => {
   try {
     const parameterId = req.params.id;
     const { parameter_name, parameter_description } = req.body;
+    const { competition_id } = req.headers;
+    if (!competition_id) {
+      return res
+        .status(500)
+        .json({ message: "Competition id not present", success: false });
+    }
 
     const query = `
       UPDATE parameters
       SET parameter_name = COALESCE($1, parameter_name), parameter_description = COALESCE($2, parameter_description)
-      WHERE parameter_id = $3
+      WHERE parameter_id = $3 AND competition_id = $4
       RETURNING *;
     `;
     //The COALESCE function is used in the UPDATE query to conditionally update only the non-null values passed in the request. This means if a value is provided in the request body, it will be updated, otherwise, the existing value will be retained.
 
-    const values = [parameter_name, parameter_description, parameterId];
+    const values = [
+      parameter_name,
+      parameter_description,
+      parameterId,
+      competition_id,
+    ];
 
     const updatedParameter = await queryDatabase(query, values);
 
@@ -97,15 +127,21 @@ export const updateParameter = async (req: Request, res: Response) => {
 
 export const deleteParameter = async (req: Request, res: Response) => {
   try {
-    const parameterId = req.params.id;
+    const parameter_id = req.params.id;
+    const { competition_id } = req.headers;
+    if (!competition_id) {
+      return res
+        .status(500)
+        .json({ message: "Competition id not present", success: false });
+    }
 
     const query = `
       DELETE FROM parameters
-      WHERE parameter_id = $1
+      WHERE parameter_id = $1 AND competition_id = $2
       RETURNING *;
     `;
 
-    const values = [parameterId];
+    const values = [parameter_id, competition_id];
 
     const deletedParameter = await queryDatabase(query, values);
